Only delete folder if existing

+++ dependency/src/main/java/de/peass/dependency/traces/ViewGeneratorThread.java
@@ -134,8 +134,9 @@ public class ViewGeneratorThread implements Runnable {
          for (final File moduleFolder : resultsManager.getExecutor().getModules().getModules()) {
             final File xmlFileFolder2 = resultsManager.getXMLFileFolder(moduleFolder);
             LOG.debug("Deleting folder: {}", xmlFileFolder2);
-            FileUtils.forceDelete(xmlFileFolder2);
-
+            if (xmlFileFolder2.exists()) {
+               FileUtils.forceDelete(xmlFileFolder2);
+            }
          }
       }
       LOG.debug("Finished: {} (Part-)Success: {}", version, gotAllData);
